requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==5.0.1
django-ckeditor==6.7.0
django-js-asset==2.2.0
Faker==22.6.0
gunicorn==21.2.0
idna==3.6
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
